remove duplicate sendJsonRpcError
