+++ src/router/AggregatorHandlers.ts
@@ -8,58 +8,7 @@ import { AggregatorService } from '../AggregatorService.js';
 import { Commitment } from '../commitment/Commitment.js';
 import logger from '../logger.js';
 import { SubmitCommitmentStatus } from '../SubmitCommitmentResponse.js';
-
-interface IJsonRpcError {
-  jsonrpc: string;
-  error: {
-    code: number;
-    message: string;
-    data?: unknown;
-  };
-  id: string | number | null;
-}
-
-/**
- * Parse a boolean value from request parameters.
- * Accepts: true, "true", 1, "1" as true
- * Accepts: false, "false", 0, "0", null, undefined as false
- */
-function parseBoolean(value: any): boolean {
-  if (typeof value === 'boolean') {
-    return value;
-  }
-  if (typeof value === 'string') {
-    return value.toLowerCase() === 'true' || value === '1';
-  }
-  if (typeof value === 'number') {
-    return value === 1;
-  }
-  return false;
-}
-
-function sendJsonRpcError(
-  res: Response,
-  httpStatus: number,
-  errorCode: number,
-  message: string,
-  id: string | number | null,
-  data?: unknown,
-): void {
-  const errorResponse: IJsonRpcError = {
-    jsonrpc: '2.0',
-    error: {
-      code: errorCode,
-      message: message,
-    },
-    id: id,
-  };
-
-  if (data !== undefined) {
-    errorResponse.error.data = data;
-  }
-
-  res.status(httpStatus).json(errorResponse);
-}
+import { sendJsonRpcError, parseBoolean } from './JsonRpcUtils.js';
 
 export async function handleSubmitCommitment(
   req: Request,
